// prisma/client/enums.ts
export const SOURCE_TYPE = {
  ARTICLE: 'ARTICLE',
  BOOK: 'BOOK',
  DIARY: 'DIARY',
  EDITION: 'EDITION',
  LETTER: 'LETTER',
  MANUSCRIPT: 'MANUSCRIPT',
  OTHER: 'OTHER'
} as const

export type SOURCE_TYPE = (typeof SOURCE_TYPE)[keyof typeof SOURCE_TYPE]
// ...
export const REVIEW_STATE = {
  PENDING: 'PENDING',
  IN_REVIEW: 'IN_REVIEW',
  APPROVED: 'APPROVED',
  ABORTED: 'ABORTED'
} as const

export type REVIEW_STATE = (typeof REVIEW_STATE)[keyof typeof REVIEW_STATE]


export const REVIEWED_ENTITY_TYPE = {
  PERSON: 'PERSON',
  ORGANIZATION: 'ORGANIZATION',
  COLLECTION: 'COLLECTION',
  PIECE: 'PIECE',
  PIECE_VERSION: 'PIECE_VERSION'
} as const

export type REVIEWED_ENTITY_TYPE = (typeof REVIEWED_ENTITY_TYPE)[keyof typeof REVIEWED_ENTITY_TYPE]


export const AUDIT_ENTITY_TYPE = {
  PERSON: 'PERSON',
  ORGANIZATION: 'ORGANIZATION',
  COLLECTION: 'COLLECTION',
  PIECE: 'PIECE',
  PIECE_VERSION: 'PIECE_VERSION',
  MOVEMENT: 'MOVEMENT',
  SECTION: 'SECTION',
  TEMPO_INDICATION: 'TEMPO_INDICATION',
  METRONOME_MARK: 'METRONOME_MARK',
  MM_SOURCE: 'MM_SOURCE',
  REFERENCE: 'REFERENCE',
  CONTRIBUTION: 'CONTRIBUTION'
} as const

export type AUDIT_ENTITY_TYPE = (typeof AUDIT_ENTITY_TYPE)[keyof typeof AUDIT_ENTITY_TYPE]


export const OPERATION = {
  CREATE: 'CREATE',
  UPDATE: 'UPDATE',
  DELETE: 'DELETE'
} as const

export type OPERATION = (typeof OPERATION)[keyof typeof OPERATION]